Refactor config initializer and options parser
This commit refactors the `ConfigInitializer` class and the `run_options_parser` method in the `options.cr` file.

In the `ConfigInitializer` class, the exception handling for creating the config file and reading the config file has been simplified. The previous code had unnecessary print statements and did not handle exceptions properly. Now, the code simply rescues any exception without printing any error messages.

In the `options.cr` file, the case statements for generating completion scripts have been modified. The previous code printed instructions on how to save the completion scripts, but those instructions have been removed. Additionally, the `banner` method in the `noir.cr` file has been removed as it was not being used.

These changes improve the readability and maintainability of the code.

Signed-off-by: HAHWUL <hahwul@gmail.com>